Backend/DataUpdater/requirements.txt
python-dotenv==1.0.0
SQLAlchemy==2.0.19
psycopg2-binary
riotwatcher==3.2.5
